requirements.txt
python-dotenv
notion-client
fastapi[standard]
pydantic
requests
schedule
pillow